+++ service/controller/clusternamespace/clusternamespace.go
@@ -1,9 +1,6 @@
 package clusternamespace
 
 import (
-	// If your operator watches a CRD import it here.
-	// "github.com/giantswarm/apiextensions/v2/pkg/apis/application/v1alpha1"
-
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"