+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use `k8smetadata` for labels and annotations instead of `apiextensions`.
+
 ## [0.23.0] - 2022-03-02
 
 ### Added